+++ cocos2d/renderer/core/pass.js
@@ -31,8 +31,9 @@ export default class Pass {
         this._depthWrite = false;
         this._depthFunc = gfx.DS_FUNC_LESS,
 
-            // stencil
-            this._stencilTest = gfx.STENCIL_INHERIT;
+        // stencil
+        this._stencilTest = gfx.STENCIL_INHERIT;
+        
         // front
         this._stencilFuncFront = gfx.DS_FUNC_ALWAYS;
         this._stencilRefFront = 0;
